Refactor GUI into package
