+++ src/prompt_automation/cli/render.py
@@ -1,34 +1,10 @@
 """Template rendering helpers for the CLI."""
 from __future__ import annotations
 
-from datetime import datetime
-from pathlib import Path
 from typing import Any
 
 from .. import menus
-
-
-def _append_to_files(var_map: dict[str, Any], text: str) -> None:
-    """Append ``text`` to any paths specified by append_file placeholders."""
-    for key, path in var_map.items():
-        if key == "append_file" or key.endswith("_append_file"):
-            if not path:
-                continue
-            try:
-                p = Path(path).expanduser()
-                with p.open("a", encoding="utf-8") as fh:
-                    if key == "context_append_file":
-                        ts = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
-                        fh.write(f"\n\n--- {ts} ---\n{text}\n")
-                    else:
-                        fh.write(text + "\n")
-            except Exception as e:
-                # Log failure silently to avoid cluttering CLI output
-                import logging
-
-                logging.getLogger("prompt_automation.cli").warning(
-                    "failed to append to %s: %s", path, e
-                )
+from ..gui.file_append import _append_to_files
 
 
 def render_template_cli(tmpl: dict[str, Any]) -> tuple[str, dict[str, Any]] | None:
