+++ src/prompt_automation/gui/__init__.py
@@ -1,0 +1,6 @@
+"""GUI package providing the :class:`PromptGUI` entry point."""
+from __future__ import annotations
+
+from .controller import PromptGUI
+
+__all__ = ["PromptGUI"]
