+++ src/prompt_automation/gui/variable_collector.py
@@ -1,0 +1,557 @@
+"""Input collection utilities for the GUI workflow."""
+from __future__ import annotations
+
+from pathlib import Path
+
+from ..renderer import read_file_safe
+from ..variables import (
+    _get_template_entry,
+    _load_overrides,
+    _print_one_time_skip_reminder,
+    _save_overrides,
+    _set_template_entry,
+)
+
+# sentinel object to signal user cancellation during input collection
+CANCELLED = object()
+
+
+def collect_file_variable_gui(template_id: int, placeholder: dict, globals_map: dict):
+    """GUI file selector with persistence and skip support."""
+    import tkinter as tk
+    from tkinter import filedialog
+
+    name = placeholder["name"]
+    label = placeholder.get("label", name)
+    # No longer support template/global skip flags via globals.json; only persisted user skip
+
+    overrides = _load_overrides()
+    entry = _get_template_entry(overrides, template_id, name) or {}
+
+    if entry.get("skip"):
+        _print_one_time_skip_reminder(overrides, template_id, name)
+        return ""
+
+    path_str = entry.get("path")
+    if path_str:
+        p = Path(path_str).expanduser()
+        if p.exists():
+            return str(p)
+        # remove stale path so user is prompted again
+        overrides.get("templates", {}).get(str(template_id), {}).pop(name, None)
+        _save_overrides(overrides)
+
+    root = tk.Tk()
+    root.title(f"File: {label}")
+    root.geometry("500x150")
+    root.resizable(False, False)
+    root.lift()
+    root.focus_force()
+    root.attributes("-topmost", True)
+    root.after(100, lambda: root.attributes("-topmost", False))
+
+    result = CANCELLED
+
+    main_frame = tk.Frame(root, padx=20, pady=20)
+    main_frame.pack(fill="both", expand=True)
+
+    path_var = tk.StringVar()
+    entry_widget = tk.Entry(main_frame, textvariable=path_var, font=("Arial", 10))
+    entry_widget.pack(side="left", fill="x", expand=True, padx=(0, 10))
+    entry_widget.focus_set()
+
+    def browse_file():
+        filename = filedialog.askopenfilename(parent=root, title=label)
+        if filename:
+            path_var.set(filename)
+
+    browse_btn = tk.Button(main_frame, text="Browse", command=browse_file, font=("Arial", 10))
+    browse_btn.pack(side="right")
+
+    button_frame = tk.Frame(root)
+    button_frame.pack(pady=(10, 0))
+
+    def on_ok():
+        nonlocal result
+        p = Path(path_var.get()).expanduser()
+        if p.exists():
+            _set_template_entry(overrides, template_id, name, {"path": str(p), "skip": False})
+            _save_overrides(overrides)
+            result = str(p)
+        else:
+            result = ""
+        root.destroy()
+
+    def on_skip():
+        nonlocal result
+        _set_template_entry(overrides, template_id, name, {"skip": True})
+        _save_overrides(overrides)
+        _print_one_time_skip_reminder(overrides, template_id, name)
+        result = ""
+        root.destroy()
+
+    def on_cancel():
+        nonlocal result
+        result = CANCELLED
+        root.destroy()
+
+    ok_btn = tk.Button(button_frame, text="OK (Enter)", command=on_ok, font=("Arial", 10), padx=20)
+    ok_btn.pack(side="left", padx=(0, 10))
+
+    skip_btn = tk.Button(button_frame, text="Skip", command=on_skip, font=("Arial", 10), padx=20)
+    skip_btn.pack(side="left", padx=(0, 10))
+
+    cancel_btn = tk.Button(button_frame, text="Cancel (Esc)", command=on_cancel, font=("Arial", 10), padx=20)
+    cancel_btn.pack(side="left")
+
+    root.bind("<Return>", lambda e: (on_ok(), "break"))
+    root.bind("<KP_Enter>", lambda e: (on_ok(), "break"))
+    root.bind("<Escape>", lambda e: (on_cancel(), "break"))
+
+    root.mainloop()
+    return result
+
+
+def collect_context_variable_gui(label: str):
+    """Collect context text with optional file loading."""
+    import tkinter as tk
+    from tkinter import filedialog
+
+    root = tk.Tk()
+    root.title(f"Context: {label}")
+    root.geometry("700x500")
+    root.resizable(True, True)
+    root.lift()
+    root.focus_force()
+    root.attributes("-topmost", True)
+    root.after(100, lambda: root.attributes("-topmost", False))
+
+    result = CANCELLED
+    selected_path = ""
+
+    main_frame = tk.Frame(root, padx=20, pady=20)
+    main_frame.pack(fill="both", expand=True)
+
+    text_frame = tk.Frame(main_frame)
+    text_frame.pack(fill="both", expand=True, pady=(0, 10))
+
+    text_widget = tk.Text(text_frame, font=("Consolas", 10), wrap="word")
+    scrollbar = tk.Scrollbar(text_frame, orient="vertical", command=text_widget.yview)
+    text_widget.config(yscrollcommand=scrollbar.set)
+    text_widget.pack(side="left", fill="both", expand=True)
+    scrollbar.pack(side="right", fill="y")
+
+    path_frame = tk.Frame(main_frame)
+    path_frame.pack(fill="x", pady=(0, 10))
+
+    path_var = tk.StringVar()
+    path_entry = tk.Entry(path_frame, textvariable=path_var, font=("Arial", 10))
+    path_entry.pack(side="left", fill="x", expand=True, padx=(0, 10))
+
+    def browse_file():
+        nonlocal selected_path
+        filename = filedialog.askopenfilename(parent=root, title=label)
+        if filename:
+            selected_path = filename
+            path_var.set(filename)
+            content = read_file_safe(filename)
+            text_widget.delete("1.0", "end")
+            text_widget.insert("1.0", content)
+
+    browse_btn = tk.Button(path_frame, text="Browse", command=browse_file, font=("Arial", 10))
+    browse_btn.pack(side="right")
+
+    button_frame = tk.Frame(main_frame)
+    button_frame.pack(pady=(10, 0))
+
+    def on_ok():
+        nonlocal result, selected_path
+        result = text_widget.get("1.0", "end-1c")
+        if path_var.get():
+            selected_path = path_var.get()
+        root.destroy()
+
+    def on_cancel():
+        nonlocal result
+        result = CANCELLED
+        root.destroy()
+
+    ok_btn = tk.Button(button_frame, text="OK (Enter)", command=on_ok, font=("Arial", 10), padx=20)
+    ok_btn.pack(side="left", padx=(0, 10))
+
+    cancel_btn = tk.Button(button_frame, text="Cancel (Esc)", command=on_cancel, font=("Arial", 10), padx=20)
+    cancel_btn.pack(side="left")
+
+    root.bind("<Control-Return>", lambda e: (on_ok(), "break"))
+    root.bind("<Control-KP_Enter>", lambda e: (on_ok(), "break"))
+    root.bind("<Escape>", lambda e: (on_cancel(), "break"))
+
+    root.mainloop()
+    return result, selected_path
+
+
+def show_reference_file_content(path: str) -> None:
+    """Display the contents of ``path`` in a read-only Tk window with light Markdown rendering."""
+    import tkinter as tk
+    import re
+
+    raw_text = read_file_safe(path)
+
+    root = tk.Tk()
+    root.title(f"Reference File: {Path(path).name}")
+    root.geometry("900x650")
+    root.resizable(True, True)
+
+    root.lift(); root.focus_force(); root.attributes("-topmost", True); root.after(100, lambda: root.attributes("-topmost", False))
+
+    main_frame = tk.Frame(root, padx=16, pady=12)
+    main_frame.pack(fill="both", expand=True)
+
+    # Toolbar
+    toolbar = tk.Frame(main_frame)
+    toolbar.pack(fill="x", pady=(0, 6))
+    info_lbl = tk.Label(
+        toolbar,
+        text="Formatted preview (basic Markdown). Toggle raw to see original.",
+        fg="#555",
+    )
+    info_lbl.pack(side="left")
+
+    text_frame = tk.Frame(main_frame)
+    text_frame.pack(fill="both", expand=True)
+
+    text_widget = tk.Text(text_frame, wrap="word", font=("Consolas", 10))
+    scrollbar = tk.Scrollbar(text_frame, orient="vertical", command=text_widget.yview)
+    text_widget.config(yscrollcommand=scrollbar.set)
+    text_widget.pack(side="left", fill="both", expand=True)
+    scrollbar.pack(side="right", fill="y")
+
+    # Tag styles
+    text_widget.tag_configure("h1", font=("Consolas", 15, "bold"))
+    text_widget.tag_configure("h2", font=("Consolas", 13, "bold"))
+    text_widget.tag_configure("h3", font=("Consolas", 12, "bold"))
+    text_widget.tag_configure("bold", font=("Consolas", 10, "bold"))
+    text_widget.tag_configure("codeblock", background="#f5f5f5", font=("Consolas", 10))
+    text_widget.tag_configure("inlinecode", background="#eee")
+    text_widget.tag_configure("hr", foreground="#666")
+
+    is_markdown = Path(path).suffix.lower() in {".md", ".markdown"} or any(
+        l.startswith("#") for l in raw_text.splitlines()[:20]
+    )
+    show_raw = {"value": False}
+
+    def render(markdown: bool):
+        text_widget.config(state="normal")
+        text_widget.delete("1.0", "end")
+        if not markdown:
+            text_widget.insert("1.0", raw_text)
+            text_widget.config(state="disabled")
+            return
+        in_code = False
+        for line in raw_text.splitlines():
+            if line.strip().startswith("```"):
+                if not in_code:
+                    in_code = True
+                    text_widget.insert("end", "\n")
+                else:
+                    in_code = False
+                continue
+            if in_code:
+                start = text_widget.index("end-1c")
+                text_widget.insert("end", line + "\n")
+                text_widget.tag_add("codeblock", start, text_widget.index("end-1c"))
+                continue
+            # Horizontal rule
+            if re.match(r"^---+$", line.strip()):
+                start = text_widget.index("end-1c")
+                text_widget.insert("end", "\n" + line.strip() + "\n")
+                text_widget.tag_add("hr", start, text_widget.index("end-1c"))
+                continue
+            # Headings
+            m = re.match(r"^(#{1,6})\s+(.*)$", line)
+            if m:
+                hashes, content = m.groups()
+                lvl = min(len(hashes), 3)
+                start = text_widget.index("end-1c")
+                text_widget.insert("end", content + "\n")
+                tag = f"h{lvl}"
+                text_widget.tag_add(tag, start, text_widget.index("end-1c"))
+                continue
+            # Bold **text**
+            pos_before = text_widget.index("end-1c")
+            working = line
+            cursor = 0
+            bold_pattern = re.compile(r"\*\*(.+?)\*\*")
+            while True:
+                bm = bold_pattern.search(working, cursor)
+                if not bm:
+                    text_widget.insert("end", working[cursor:] + "\n")
+                    break
+                # text before bold
+                text_widget.insert("end", working[cursor:bm.start()])
+                bold_start = text_widget.index("end-1c")
+                text_widget.insert("end", bm.group(1))
+                bold_end = text_widget.index("end-1c")
+                text_widget.tag_add("bold", bold_start, bold_end)
+                cursor = bm.end()
+                if cursor >= len(working):
+                    text_widget.insert("end", "\n")
+                    break
+            # Inline code `code`
+            # After insertion, scan the just inserted line region for backticks
+            line_start_idx = pos_before
+            line_end_idx = text_widget.index("end-2c lineend")
+            line_text = text_widget.get(line_start_idx, line_end_idx)
+            inline_positions = [
+                (m.start(), m.end()) for m in re.finditer(r"`([^`]+)`", line_text)
+            ]
+            # Adjust tags; remove backticks visually by replacing them
+            for s, e in reversed(inline_positions):
+                # Replace with inner content
+                inner = re.match(r"`([^`]+)`", line_text[s:e]).group(1)
+                abs_start = f"{line_start_idx.split('.')[0]}.{int(line_start_idx.split('.')[1]) + s}"
+                abs_end = f"{line_start_idx.split('.')[0]}.{int(line_start_idx.split('.')[1]) + e}"
+                text_widget.delete(abs_start, abs_end)
+                text_widget.insert(abs_start, inner)
+                # Compute new end after replacement
+                new_end_idx = f"{line_start_idx.split('.')[0]}.{int(line_start_idx.split('.')[1]) + s + len(inner)}"
+                text_widget.tag_add("inlinecode", abs_start, new_end_idx)
+        text_widget.config(state="disabled")
+
+    def toggle_view():
+        show_raw["value"] = not show_raw["value"]
+        raw_btn.config(text="Raw View" if not show_raw["value"] else "Formatted View")
+        render(markdown=is_markdown and not show_raw["value"])
+
+    def copy_all():
+        try:
+            import prompt_automation.paste as paste
+
+            paste.copy_to_clipboard(raw_text)
+        except Exception:
+            pass
+
+    raw_btn = tk.Button(toolbar, text="Raw View", command=toggle_view)
+    raw_btn.pack(side="right")
+    copy_btn = tk.Button(toolbar, text="Copy All", command=copy_all)
+    copy_btn.pack(side="right", padx=(0, 6))
+
+    button_frame = tk.Frame(main_frame)
+    button_frame.pack(pady=(8, 0))
+
+    def on_close(event=None):  # pragma: no cover - GUI
+        root.destroy()
+        return "break"
+
+    close_btn = tk.Button(button_frame, text="Close (Esc)", command=on_close, font=("Arial", 10), padx=20)
+    close_btn.pack()
+    root.bind("<Escape>", on_close)
+    root.bind("<Return>", on_close)
+    root.mainloop()
+
+
+def collect_variables_gui(template):
+    """Collect variables for template placeholders - fully keyboard driven."""
+    placeholders = template.get("placeholders", [])
+    if not placeholders:
+        return {}
+
+    variables = {}
+    template_id = template.get("id", 0)
+    globals_map = template.get("global_placeholders", {})
+
+    for placeholder in placeholders:
+        name = placeholder["name"]
+        label = placeholder.get("label", name)
+        ptype = placeholder.get("type", "text")
+        options = placeholder.get("options", [])
+        multiline = placeholder.get("multiline", False) or ptype == "list"
+
+        if name == "reference_file_content":
+            path = variables.get("reference_file")
+            p = Path(path).expanduser() if path else None
+            if p and p.exists():
+                show_reference_file_content(str(p))
+                variables[name] = read_file_safe(str(p))
+            else:
+                variables[name] = ""
+            continue
+
+        if name == "context":
+            value, ctx_path = collect_context_variable_gui(label)
+            if value is CANCELLED:
+                return None
+            variables[name] = value
+            if ctx_path:
+                variables["context_append_file"] = ctx_path
+            continue
+
+        if ptype == "file":
+            value = collect_file_variable_gui(template_id, placeholder, globals_map)
+        else:
+            value = collect_single_variable(name, label, ptype, options, multiline)
+        if value is CANCELLED:  # User cancelled entire workflow
+            return None
+        variables[name] = value
+
+    return variables
+
+
+def collect_single_variable(name, label, ptype, options, multiline):
+    """Collect a single variable with appropriate input method."""
+    import tkinter as tk
+    from tkinter import ttk, filedialog
+
+    root = tk.Tk()
+    root.title(f"Input: {label}")
+    root.geometry("500x300" if multiline else "500x150")
+    root.resizable(False, False)
+
+    # Bring to foreground and focus
+    root.lift()
+    root.focus_force()
+    root.attributes("-topmost", True)
+    root.after(100, lambda: root.attributes("-topmost", False))
+
+    result = CANCELLED
+
+    # Main frame
+    main_frame = tk.Frame(root, padx=20, pady=20)
+    main_frame.pack(fill="both", expand=True)
+
+    # Label
+    label_widget = tk.Label(main_frame, text=f"{label}:", font=("Arial", 12))
+    label_widget.pack(anchor="w", pady=(0, 10))
+
+    # Input widget based on type
+    input_widget = None
+
+    if options:
+        # Dropdown for options
+        input_widget = ttk.Combobox(main_frame, values=options, font=("Arial", 10))
+        input_widget.pack(fill="x", pady=(0, 10))
+        input_widget.set(options[0] if options else "")
+        input_widget.focus_set()
+
+    elif ptype == "file":
+        # File input with browse button
+        file_frame = tk.Frame(main_frame)
+        file_frame.pack(fill="x", pady=(0, 10))
+
+        input_widget = tk.Entry(file_frame, font=("Arial", 10))
+        input_widget.pack(side="left", fill="x", expand=True, padx=(0, 10))
+
+        def browse_file():
+            filename = filedialog.askopenfilename(parent=root)
+            if filename:
+                input_widget.delete(0, "end")
+                input_widget.insert(0, filename)
+
+        browse_btn = tk.Button(
+            file_frame,
+            text="Browse",
+            command=browse_file,
+            font=("Arial", 10),
+        )
+        browse_btn.pack(side="right")
+
+        input_widget.focus_set()
+
+    elif multiline or ptype == "list":
+        # Multi-line text input
+        text_frame = tk.Frame(main_frame)
+        text_frame.pack(fill="both", expand=True, pady=(0, 10))
+
+        input_widget = tk.Text(text_frame, font=("Arial", 10), wrap="word")
+        scrollbar = tk.Scrollbar(text_frame, orient="vertical", command=input_widget.yview)
+        input_widget.config(yscrollcommand=scrollbar.set)
+
+        input_widget.pack(side="left", fill="both", expand=True)
+        scrollbar.pack(side="right", fill="y")
+
+        input_widget.focus_set()
+
+    else:
+        # Single-line text input
+        input_widget = tk.Entry(main_frame, font=("Arial", 10))
+        input_widget.pack(fill="x", pady=(0, 10))
+        input_widget.focus_set()
+
+    # Button frame
+    button_frame = tk.Frame(main_frame)
+    button_frame.pack(fill="x")
+
+    def on_ok(skip: bool = False):
+        nonlocal result
+        if skip:
+            result = None
+        elif isinstance(input_widget, tk.Text):
+            value = input_widget.get("1.0", "end-1c")
+            if ptype == "list":
+                result = [line.strip() for line in value.splitlines() if line.strip()]
+            else:
+                result = value
+        else:
+            result = input_widget.get()
+        root.destroy()
+
+    def on_cancel():
+        nonlocal result
+        result = CANCELLED
+        root.destroy()
+
+    ok_btn = tk.Button(button_frame, text="OK (Enter)", command=on_ok, font=("Arial", 10), padx=20)
+    ok_btn.pack(side="left", padx=(0, 10))
+
+    cancel_btn = tk.Button(
+        button_frame,
+        text="Cancel (Esc)",
+        command=on_cancel,
+        font=("Arial", 10),
+        padx=20,
+    )
+    cancel_btn.pack(side="left")
+
+    # Keyboard bindings
+    def on_enter(event):
+        # For multi-line text, Ctrl+Enter submits, Enter adds new line
+        is_ctrl = bool(event.state & 0x4)
+        if isinstance(input_widget, tk.Text) and not is_ctrl:
+            return None  # Allow normal Enter behavior in text widget
+
+        if is_ctrl:
+            if isinstance(input_widget, tk.Text):
+                current = input_widget.get("1.0", "end-1c").strip()
+            else:
+                current = input_widget.get().strip()
+            if not current:
+                on_ok(skip=True)
+                return "break"
+        on_ok()
+        return "break"
+
+    def on_escape(event):
+        on_cancel()
+        return "break"
+
+    root.bind("<Control-Return>", on_enter)
+    root.bind("<Control-KP_Enter>", on_enter)
+    root.bind("<Escape>", on_escape)
+
+    # For non-text widgets, regular Enter also submits
+    if not isinstance(input_widget, tk.Text):
+        root.bind("<Return>", on_enter)
+        root.bind("<KP_Enter>", on_enter)
+
+    root.mainloop()
+    return result
+
+
+__all__ = [
+    "CANCELLED",
+    "collect_file_variable_gui",
+    "collect_context_variable_gui",
+    "show_reference_file_content",
+    "collect_variables_gui",
+    "collect_single_variable",
+]
